Restructured file structure and removed unnecessary imports from routes
