+++ Service/src/auth/login.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, request, Response, jsonify, g
 from flask_login import LoginManager
 from werkzeug.security import generate_password_hash, check_password_hash
-import manageTokens
+from auth import manageTokens
 from util import logger
 
 log = logger.Logger(True, True, True)
