+++ Service/src/routes/league.py
@@ -1,19 +1,18 @@
 from flask import Blueprint, request, Response, g, jsonify
-from util import logger
-from datetime import datetime
-import pymysql
 import simplejson as json
-import random
-import string
+
+from util import logger
+from util.utility import Utility
 
 log = logger.Logger(True, True, True)
 
 league_api = Blueprint('league_api', __name__)
 
+
 @league_api.route('/api/league', methods=['POST'])
 def post_league():
    body = request.get_json()
-   inviteCode = gen_inviteCode()
+   inviteCode = Utility.gen_inviteCode()
 
    g.cursor.execute("INSERT INTO League(name, start, end, startPos, inviteCode, ownerId) VALUES (%s, %s, %s, %s, %s, %s)",
       [body['name'], body['start'], body['end'], body['startPos'], inviteCode, body['ownerId']])
@@ -26,16 +25,4 @@ def get_leagues():
    g.cursor.execute("SELECT * FROM League")
 
    leagues = g.cursor.fetchall()
-   return json.dumps(leagues, default=dateToStr)
-
-
-# TODO move to util folder
-def gen_inviteCode():
-   code = random.sample(string.ascii_uppercase + string.digits, 6)
-   return ''.join(code)
-
-
-# TODO move to util folder
-def dateToStr(obj):
-   if isinstance(obj, datetime):
-      return obj.__str__()
+   return json.dumps(leagues, default=Utility.dateToStr)
