+++ Service/src/util/utility.py
@@ -1,0 +1,14 @@
+from datetime import datetime
+import random
+import string
+
+class Utility:
+
+   def gen_inviteCode():
+      code = random.sample(string.ascii_uppercase + string.digits, 6)
+      return ''.join(code)
+
+
+   def dateToStr(obj):
+      if isinstance(obj, datetime):
+         return obj.__str__()
